refactor(evaluations): TakeEvaluation sous 300 lignes — éclatement complet (H1)
Lignes: TakeEvaluation.vue 1014 → 150.

Fichiers créés:
- src/composables/useTakeEvaluation.js (227) — tout le <script> (Options API → Composition API): état, computed (answeredCount, progressPercentage), timers (temps imparti + fenêtre temporelle), chargement, soumission, résultats. Services/routes/payloads/alert/confirm/console.log identiques.
- src/components/evaluations/TakeEvaluationBanners.vue (150) — bandeau entraînement + alertes fenêtre.
- src/components/evaluations/TakeEvaluationHeader.vue (162) — titre/minuteur + barre de progression.
- src/components/evaluations/TakeQuestionCard.vue (227) — carte question (5 types: qcm, qcm_multiple, vrai_faux, reponse_courte, dissertation).
- src/components/evaluations/TakeEvaluationActions.vue (109) — pied d'actions Annuler/Soumettre.
- src/components/evaluations/TakeResultModals.vue (235) — modales confirmation + résultats.
- tests/unit/useTakeEvaluation.test.js (137), tests/unit/TakeComponents.test.js (135).

Orchestrateur: coque layout + état d'erreur + CSS résiduel (.take-evaluation-container, .questions-list, .error-*), câble useTakeEvaluation et les sous-composants.

CSS déplacé verbatim (keyframe pulse + .btn-cancel/.btn-submit dupliqués car partagés entre sections).

Tests: 17/17 verts. Build: vite build exit 0 (warning chunks >500kB préexistant).
Parité stricte: mêmes classes/textes/services/routes/payloads, comportement identique.
